pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "korean-news-collector"
version = "0.2.0"
description = "구글 뉴스 수집기 - Tavily 검색 + Playwright 본문 추출"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "배움의 달인", email = "learn@example.com" }]
keywords = ["news", "korean", "tavily", "playwright", "scraper"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Internet :: WWW/HTTP",
]
dependencies = [
    "tavily-python>=0.5.0",
    "playwright>=1.40.0",
    "beautifulsoup4>=4.12.0",
    "httpx>=0.27.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "lxml>=4.9.0",
]

[project.scripts]
knews = "knews.cli:main"

[project.urls]
Homepage = "https://github.com/daniel8824/korean-news-collector"
Repository = "https://github.com/daniel8824/korean-news-collector"

[tool.hatch.build.targets.wheel]
packages = ["src/knews"]

[tool.hatch.build.targets.sdist]
include = ["src/knews"]
